+++ migrations/0030_heavy_leper_queen.sql
@@ -1,0 +1,4 @@
+ALTER TABLE "plans" ADD COLUMN "custom_layouts_enabled" boolean DEFAULT true NOT NULL;--> statement-breakpoint
+ALTER TABLE "plans" ADD COLUMN "playlists_enabled" boolean DEFAULT true NOT NULL;--> statement-breakpoint
+ALTER TABLE "plans" ADD COLUMN "remote_control_enabled" boolean DEFAULT true NOT NULL;--> statement-breakpoint
+ALTER TABLE "plans" ADD COLUMN "max_projects" integer;